Has a working drive arm speed...Does not have the telem task or systemprint active

+++ Software/CCS/RoverMotherboard/roveIncludes/roveWareHeaders/roveCmdCntrl.h
@@ -38,8 +38,8 @@
 
 Void roveCmdCntrl(UArg arg0, UArg arg1);
 
-void roboArmPositiveWrite(int struct_id, int speed, char* output_buffer);
+void roboArmPositiveWrite(int struct_id, int16_t speed, char* output_buffer);
 
-void roboArmNegativeWrite(int struct_id, int speed, char* output_buffer);
+void roboArmNegativeWrite(int struct_id, int16_t speed, char* output_buffer);
 
 #endif //ROVECMDCNTRL_H_
